requirements.txt
paho-mqtt==1.6.1
python-dotenv==0.19.2
requests==2.27.1
